Start removing redux
