+++ .gitignore
@@ -2,3 +2,4 @@
 lib
 node_modules
 package-lock.json
+.DS_Store
