--- src/Signer/ConfirmViaKey/confirmViaKey.css
@@ -1,48 +1,0 @@
-/* Copyright 2015-2017 Parity Technologies (UK) Ltd.
-/* This file is part of Parity.
-/*
-/* Parity is free software: you can redistribute it and/or modify
-/* it under the terms of the GNU General Public License as published by
-/* the Free Software Foundation, either version 3 of the License, or
-/* (at your option) any later version.
-/*
-/* Parity is distributed in the hope that it will be useful,
-/* but WITHOUT ANY WARRANTY; without even the implied warranty of
-/* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-/* GNU General Public License for more details.
-/*
-/* You should have received a copy of the GNU General Public License
-/* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
-*/
-.confirmForm {
-  margin-top: -2em;
-}
-
-.confirmButton {
-  display: block !important;
-  margin-bottom: 10px;
-  white-space: nowrap;
-}
-
-.signerIcon {
-  vertical-align: middle;
-}
-
-.passwordHint {
-  font-size: 0.75em;
-  color: #aaa;
-  margin-bottom: 0.75em;
-}
-
-.passwordHint span {
-  opacity: 0.85;
-}
-
-.fileInput input {
-  top: 22px;
-}
-
-.qr {
-  margin-bottom: 0.5em;
-  text-align: center;
-}
