+++ src/Signer/Request/RequestSign/requestSign.css
@@ -15,7 +15,7 @@
 /* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-@import '../_layout';
+@import '../../_layout';
 
 .container {
   display: flex;
